Split out Bounceable, move to components folder
