+++ client/src/components/MobileSimulator.css
